BinningFactory now skips null values when analysing values (NPE before).  Also removed option to specify the number of bins when sorting values, resulting in always having 'enough' values to create equal count bins, which prevented some of the binning code intended to handle that situation from ever seeing low value counts.
Some of the BinningFactoryTests are broken from previous changes related to how infinate vavlues.  I added a TODO which will need to be resolved once we determine how to handle this.

Added a new Bug1 test that replicates the data from a session Lorraine M. found that creates really ugly looking bins.  This test currently fails, but I want it checked in over the weekend.


Former-commit-id: e6543444ec582e6af54318b3aeccb7e5994b1212