issue SPDSS-805: Make changes to the components of the dss to look at the views that share geometry between models http://internal.cida.usgs.gov/jira/browse/SPDSS-805
Model metadata now include a themeName attribute (store in the db as THEME_NAME in the SPARROW_MODEL table) that indicates which MapViewer predefined theme should be used for the model.  'reach_' or 'catch_' is appended to the front of the theme name to match the actual predefined theme names mapviewer uses.

* MapRequestServlet had some old params related to rendering calibrations sites removed (now handled in a separate servlet)
* MapRequestServlet cleanup and added param filtering.
* SparrowUIContext now has a transient themeName property.


Former-commit-id: 2638f39030999f9ba21b80d87894e9be86dc71e1