issue SPDSS-819: Under heavy load, Oracle cursors are left open for the SPARROW_DSS_TRANS.GENERIC_CACHE_BY_CLS_AND_CODE procedure http://internal.cida.usgs.gov/jira/browse/SPDSS-819
I found Oracle notes stating that ResultSets are not closed when the statement is closed.  Generally I've seen this to *not* be true - Statements will close their associated ResultSets.  However, in the case of the Generic cache, we retrieve the resultSet as a return parameter, which is a bit weird.  As a result, its possible that it is not being auto-closed w/ the statement.  Now updated all actions to auto-close the resultset.


Former-commit-id: 8d26b32fcc19a3767b4cd2c4d2f5c255107f8ae5