issue SPDSS-903: Display only relevant stream network links and info in export header http://internal.cida.usgs.gov/jira/browse/SPDSS-903
Previous version of the streaming export use simply place the readme notes in the 'start document delimiter'.  That solution worked until this JIRA, which requires the readme to be dynamic.  However, the formatter code is unaware of model details, so it just injects the header as static text.

The solution was to add the readme text as a full-fledge <description> element in the document, where it can be read and handled appropriately by the formatter.  Since the description element is 'real' content, it can be made dynamic just like the rest of the document.

* BasicTagEven now supports CDATA events, needed to contain the ad-hoc content of the readme.
* SparrowModel and related classes now loads and contains the appropriate 'enhanced' network fields from the db to populate the readme file.
* SparrowResourceUtils now includes lookupMergedHelp method that does string replacement.
* To allow multi-row contents in Excel (ie, a carriage return in a cell), the excel styles were extended to include a style that allows text wrapping.


Former-commit-id: 700e5ebae3cbb3d17f67d951efb865f98fc33e8e