SPDSS-1021
Agg Delivery Report may fail in some cases

The info table always has one reach for each region in the model (states in this case).
The data table (values for each area) should have the same number of rows.
Later these rows are filtered to remove the zero value rows.

The issue here was that, since the data table is build dynamically, its columns do not add values beyond where values were written to. For instance, if the last state is West Virginia and no value is written for that state (ie, there are no reaches upstream of the terminal reaches in W. Virginia), then the table rows will not extend to the last row of the table.

Fixed by writing a zero value to the data table, corresponding to the row of the last row of the region table.


Former-commit-id: 3ba1c519c00a906525f1ba02efa06e574a8a65e5