Major revamp of CalcReachAreaFractionMap, CalcFractionedWatershedAre, and PredictData topo data.  PredictData now returns a TopoData object for getTopo (instead of just a DataTable).  The TopoData object contains built-in topological operations on its data.  These operations are used extensively in the CalcReachAreaFractionMap action.
CalcReachAreaFractionMap now attempts to use an optimized calculation where it checks to see if the immediately upstream reach areas have already been calculated.  This hugely speeds up large operations if done in the correct order.  For this to work, the size of the ehcache for calculated watershed areas was increased.

Several classes were affected by the change in the topo return type.

* New optimized FractionalReach to store just the info needed for an in-process reach in CalcReachAreaFractionMap.
* Add new tests for CalcReachAreaFractionMap to test braided streams.



Former-commit-id: de8cc93ed364ae29d5985ec845581e1dad4b3340